replace inputloopservice
