+++ samples/routing/command-routing/Core_10/Sender/Program.cs
@@ -1,13 +1,12 @@
 using System;
+using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
-using Sender;
 
 Console.Title = "Sender";
 
 var builder = Host.CreateApplicationBuilder(args);
-builder.Services.AddHostedService<InputLoopService>();
 
 var endpointConfiguration = new EndpointConfiguration("Samples.CommandRouting.Sender");
 
@@ -22,10 +21,60 @@ routing.RouteToEndpoint(
 
 endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 
+builder.UseNServiceBus(endpointConfiguration);
+
+var host = builder.Build();
 
 Console.WriteLine("Press any key, the application is starting");
 Console.ReadKey();
 Console.WriteLine("Starting...");
 
-builder.UseNServiceBus(endpointConfiguration);
-await builder.Build().RunAsync();
+await host.StartAsync();
+var messageSession = host.Services.GetRequiredService<IMessageSession>();
+
+Console.WriteLine("Press S to send an order");
+Console.WriteLine("Press C to cancel an order");
+Console.WriteLine("Press ESC to exit");
+
+ConsoleKey keyPressed = Console.ReadKey(true).Key;
+
+while (keyPressed != ConsoleKey.Escape)
+{
+    switch (keyPressed)
+    {
+        case ConsoleKey.S:
+            await PlaceOrder(messageSession, Guid.NewGuid().ToString(), 25m);
+            break;
+        case ConsoleKey.C:
+            await CancelOrder(messageSession, Guid.NewGuid().ToString());
+            break;
+    }
+
+    keyPressed = Console.ReadKey(true).Key;
+}
+
+await host.StopAsync();
+
+static async Task PlaceOrder(IMessageSession messageSession, string orderId, decimal value)
+{
+    var command = new PlaceOrder
+    {
+        OrderId = orderId,
+        Value = value
+    };
+
+    await messageSession.Send(command);
+    Console.WriteLine($"Order placed: {orderId}");
+}
+
+static async Task CancelOrder(IMessageSession messageSession, string orderId)
+{
+    var command = new CancelOrder
+    {
+        OrderId = orderId
+    };
+
+    await messageSession.Send("Samples.CommandRouting.Receiver", command);
+    Console.WriteLine($"Order canceled: {orderId}");
+}
+
